Taking out scale variable
